Implement fade out/fade in

+++ include/OpenGL/Iris2D/ViewportGL.h
@@ -43,7 +43,7 @@ namespace Iris2D {
 		float m_fY = 0.0f;
 		float m_fZ = 0.0f;
 
-		std::multimap<unsigned int, SpriteGL*, std::less<>> m_stSprites;
+		std::multimap<float, SpriteGL*, std::less<>> m_stSprites;
 
 		unsigned int m_nVAO = 0;
 		unsigned int m_nVBO = 0;
